requirements.txt
# # Core FastAPI dependencies
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# pydantic==2.5.3
# pydantic-settings==2.1.0

# # Vector Database
# chromadb==0.4.22

# # Embeddings and ML
# sentence-transformers==2.3.1
# # torch==2.1.2
# torch>=2.0.0

# # PDF Processing
# pypdf==4.0.1

# # LLM Integration
# google-generativeai==0.3.2

# # Utilities
# python-multipart==0.0.6
# python-dotenv==1.0.0
# aiofiles==23.2.1

# # Development
# pytest==7.4.4
# httpx==0.26.0

# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.3.0

# Vector Database
chromadb>=0.4.22

# Embeddings and ML
sentence-transformers>=2.3.0
torch>=2.0.0

# PDF Processing
pypdf>=4.0.0

# LLM Integration
google-generativeai>=0.3.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Development
pytest>=7.4.0
httpx>=0.26.0